addd

+++ ft_strlen.c
@@ -1,9 +1,8 @@
-
 #include "libft.h"
 
 size_t	ft_strlen(const char *str)
 {
-	int len;
+	size_t	len;
 
 	len = 0;
 	while (str[len])
